Move WaitForDependencies into the ResourceNotificationService.
