+++ src/Aspire.Hosting.Azure/Provisioning/Provisioners/AzureProvisioner.cs
@@ -10,7 +10,6 @@ using Aspire.Hosting.Azure.Provisioning;
 using Aspire.Hosting.Azure.Utils;
 using Aspire.Hosting.Eventing;
 using Aspire.Hosting.Lifecycle;
-using Aspire.Hosting.Orchestration;
 using Azure;
 using Azure.Core;
 using Azure.Identity;
@@ -35,8 +34,7 @@ internal sealed class AzureProvisioner(
     IServiceProvider serviceProvider,
     ResourceNotificationService notificationService,
     ResourceLoggerService loggerService,
-    IDistributedApplicationEventing eventing,
-    IDistributedApplicationOrchestrator orchestrator
+    IDistributedApplicationEventing eventing
     ) : IDistributedApplicationLifecycleHook
 {
     internal const string AspireResourceNameTag = "aspire-resource-name";
@@ -240,8 +238,6 @@ internal sealed class AzureProvisioner(
 
     private async Task ProcessResourceAsync(IConfiguration configuration, Lazy<Task<ProvisioningContext>> provisioningContextLazy, (IResource Resource, IAzureResource AzureResource) resource, CancellationToken cancellationToken)
     {
-        await orchestrator.WaitForDependenciesAsync(resource.Resource, cancellationToken).ConfigureAwait(false);
-
         var beforeResourceStartedEvent = new BeforeResourceStartedEvent(resource.Resource, serviceProvider);
         await eventing.PublishAsync(beforeResourceStartedEvent, cancellationToken).ConfigureAwait(false);
